Refactor: Consolidate and rename components for clarity
This commit reorganizes and renames several components to improve code structure and maintainability. The 'About' component is now 'AboutCompany', 'Benefits' is merged into 'InvestmentBenefits', and 'Infrastructure' and 'TechSolutions' are combined into a new 'Technology' component. This consolidation aims to streamline the codebase and reduce redundancy.
