+++ app/page.jsx
@@ -1,12 +1,9 @@
 import Header from '@/components/Header';
 import Hero from '@/components/Hero';
-import About from '@/components/About';
-import Organogram from '@/components/Organogram';
+import AboutCompany from '@/components/AboutCompany';
 import Services from '@/components/Services';
-import TechSolutions from '@/components/TechSolutions';
-import Infrastructure from '@/components/Infrastructure';
-import Investment from '@/components/Investment';
-import Benefits from '@/components/Benefits';
+import Technology from '@/components/Technology';
+import InvestmentBenefits from '@/components/InvestmentBenefits';
 import Contact from '@/components/Contact';
 import Footer from '@/components/Footer';
 
@@ -16,13 +13,10 @@ export default function Home() {
       <Header />
       <main>
         <Hero />
-        <About />
-        <Organogram />
+        <AboutCompany />
         <Services />
-        <TechSolutions />
-        <Infrastructure />
-        <Investment />
-        <Benefits />
+        <Technology />
+        <InvestmentBenefits />
         <Contact />
       </main>
       <Footer />
